[refactor] schema.sql 초기 데이터 설정, 패키지 구조 변경
