+++ src/main/java/gift/feat/product/repository/ProductRepository.java
@@ -1,8 +1,8 @@
-package gift.repository;
+package gift.feat.product.repository;
 
 import java.util.List;
 import java.util.Optional;
-import gift.domain.Product;
+import gift.feat.product.domain.Product;
 
 public interface ProductRepository {
 	void save(Product product);
