+++ src/main/resources/templates/admin/productPage.html
